Added data validation

+++ backend/Controllers/CommentController.cs
@@ -14,51 +14,60 @@ namespace backend.Controllers
     public class CommentController : ControllerBase
     {
         private readonly ICommentRepository _commentRepo;
-        public CommentController(ICommentRepository commentRepo)
+        private readonly IStockRepository _stockRepo;
+        public CommentController(ICommentRepository commentRepo, IStockRepository stockRepo)
         {
             _commentRepo = commentRepo;
+            _stockRepo = stockRepo;
         }
-        
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid) return BadRequest(ModelState);
             var comments = await _commentRepo.GetAllAsync();
-            var commentDto = comments.Select(s=>s.ToCommentDto());
+            var commentDto = comments.Select(s => s.ToCommentDto());
             return Ok(comments);
         }
-
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var comment = await _commentRepo.GetByIdAsync(id);
-            if(comment == null) return NotFound();
+            if (comment == null) return NotFound();
             return Ok(comment.ToCommentDto());
         }
-
-        [HttpPost]
-        public async Task<IActionResult> Create([FromBody] CreateCommentRequestDto commentDto)
+        [HttpPost("{stockId:int}")]
+        public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentRequestDto commentDto)
         {
-            var commentModel = commentDto.ToCommentFromCreateDTO();
-            await _commentRepo.CreateAsync(commentModel);
-            return CreatedAtAction(nameof(GetById), 
-                new{ id = commentModel.Id}, commentModel.ToCommentDto());
-        }
+            if(!await _stockRepo.StockExists(stockId)) return BadRequest("Stock does not exist");
+            if (!ModelState.IsValid) return BadRequest(ModelState);
 
+
+            var commentModel = commentDto.ToCommentFromCreateDTO(stockId);
+            await _commentRepo.CreateAsync(commentModel);
+            return CreatedAtAction(nameof(GetById),
+                new { id = commentModel }, commentModel.ToCommentDto());
+        }
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var commentModel = await _commentRepo.UpdateAsync(id, updateDto);
-            if(commentModel == null) return NotFound();
+            if (commentModel == null) return NotFound();
             return Ok(commentModel.ToCommentDto());
         }
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
+
             var commentModel = await _commentRepo.DeleteAsync(id);
-            if(commentModel == null) return NotFound();
-            return NoContent();
+            if (commentModel == null) return NotFound("Comment does not exist");
+            return Ok(commentModel);
         }
     }
 }
